Refaktorovan kod nakon promene imena polja u RADE API-u.

+++ be_t4_04_project/src/main/java/com/ikt/t4/project/repositories/OfferRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.ikt.t4.project.entities.OfferEntity;
 
-public interface OfferRepository extends CrudRepository<OfferEntity, Integer> {
+public interface OfferRepository extends CrudRepository<OfferEntity, Long> {
 
 }
